--- python/cusignal/cupy_helper/linalg_extended.py
@@ -1,122 +1,0 @@
-# Copyright (c) 2019-2020, NVIDIA CORPORATION.
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import cupy as cp
-from cupy.lib.stride_tricks import as_strided
-
-
-def toeplitz(c, r=None):
-    """
-    Construct a Toeplitz matrix.
-    The Toeplitz matrix has constant diagonals, with c as its first column
-    and r as its first row.  If r is not given, ``r == conjugate(c)`` is
-    assumed.
-    Parameters
-    ----------
-    c : array_like
-        First column of the matrix.  Whatever the actual shape of `c`, it
-        will be converted to a 1-D array.
-    r : array_like, optional
-        First row of the matrix. If None, ``r = conjugate(c)`` is assumed;
-        in this case, if c[0] is real, the result is a Hermitian matrix.
-        r[0] is ignored; the first row of the returned matrix is
-        ``[c[0], r[1:]]``.  Whatever the actual shape of `r`, it will be
-        converted to a 1-D array.
-    Returns
-    -------
-    A : (len(c), len(r)) ndarray
-        The Toeplitz matrix. Dtype is the same as ``(c[0] + r[0]).dtype``.
-    See Also
-    --------
-    circulant : circulant matrix
-    hankel : Hankel matrix
-    solve_toeplitz : Solve a Toeplitz system.
-    Notes
-    -----
-    The behavior when `c` or `r` is a scalar, or when `c` is complex and
-    `r` is None, was changed in version 0.8.0.  The behavior in previous
-    versions was undocumented and is no longer supported.
-    Examples
-    --------
-    >>> from scipy.linalg import toeplitz
-    >>> toeplitz([1,2,3], [1,4,5,6])
-    array([[1, 4, 5, 6],
-           [2, 1, 4, 5],
-           [3, 2, 1, 4]])
-    >>> toeplitz([1.0, 2+3j, 4-1j])
-    array([[ 1.+0.j,  2.-3.j,  4.+1.j],
-           [ 2.+3.j,  1.+0.j,  2.-3.j],
-           [ 4.-1.j,  2.+3.j,  1.+0.j]])
-    """
-    c = cp.asarray(c).ravel()
-    if r is None:
-        r = cp.conjugate()
-    else:
-        r = cp.asarray(r).ravel()
-    # Form a 1D array containing a reversed c followed by r[1:] that could be
-    # strided to give us toeplitz matrix.
-    vals = cp.concatenate((c[::-1], r[1:]))
-    out_shp = len(c), len(r)
-    n = vals.strides[0]
-    return as_strided(vals[len(c) - 1:], shape=out_shp, strides=(-n, n)).copy()
-
-
-def hankel(c, r=None):
-    """
-    Construct a Hankel matrix.
-    The Hankel matrix has constant anti-diagonals, with `c` as its
-    first column and `r` as its last row.  If `r` is not given, then
-    `r = zeros_like(c)` is assumed.
-    Parameters
-    ----------
-    c : array_like
-        First column of the matrix.  Whatever the actual shape of `c`, it
-        will be converted to a 1-D array.
-    r : array_like, optional
-        Last row of the matrix. If None, ``r = zeros_like(c)`` is assumed.
-        r[0] is ignored; the last row of the returned matrix is
-        ``[c[-1], r[1:]]``.  Whatever the actual shape of `r`, it will be
-        converted to a 1-D array.
-    Returns
-    -------
-    A : (len(c), len(r)) ndarray
-        The Hankel matrix. Dtype is the same as ``(c[0] + r[0]).dtype``.
-    See Also
-    --------
-    toeplitz : Toeplitz matrix
-    circulant : circulant matrix
-    Examples
-    --------
-    >>> from scipy.linalg import hankel
-    >>> hankel([1, 17, 99])
-    array([[ 1, 17, 99],
-           [17, 99,  0],
-           [99,  0,  0]])
-    >>> hankel([1,2,3,4], [4,7,7,8,9])
-    array([[1, 2, 3, 4, 7],
-           [2, 3, 4, 7, 7],
-           [3, 4, 7, 7, 8],
-           [4, 7, 7, 8, 9]])
-    """
-    c = cp.asarray(c).ravel()
-    if r is None:
-        r = cp.zeros_like(c)
-    else:
-        r = cp.asarray(r).ravel()
-    # Form a 1D array of values to be used in the matrix, containing `c`
-    # followed by r[1:].
-    vals = cp.concatenate((c, r[1:]))
-    # Stride on concatenated array to get hankel matrix
-    out_shp = len(c), len(r)
-    n = vals.strides[0]
-    return as_strided(vals, shape=out_shp, strides=(n, n)).copy()
